drivers/soc/cse: Fix overflow in CSE telemetry calculation
MSEC_TO_USEC(cse_perf_data.timestamp[i]) does overflow.
Here is an example, if cse_perf_data.timestamp[i] value is 4304903 milliseconds.
When multiplied by 1000 to convert to microseconds, the value becomes
0x979B58 instead of 0x100979B58.

TEST=modify cse_perf_data.timestamp values and make sure they don't
overflow.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6210782
Tested-by: Subrata Banik <subratabanik@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>

+++ src/drivers/soc/cse/cse_telemetry.c
@@ -16,7 +16,7 @@
 #include "cse.h"
 #include "cse_telemetry.h"
 
-#define MSEC_TO_USEC(x) (x * 1000)
+#define MSEC_TO_USEC(x) ((s64)x * 1000)
 
 #if CONFIG(SOC_INTEL_CSE_PRE_CPU_RESET_TELEMETRY_V1)
 static int soc_cbmem_inject_telemetry_data(s64 *time_stamp, s64 current_time)
